chore: add ui subset barrels
