+++ frontends/nextjs/src/components/ui/core.ts
@@ -1,0 +1,87 @@
+// Core UI exports: foundational atoms and common building blocks
+export * from './atoms'
+
+// Common molecules
+export {
+  Card,
+  CardHeader,
+  CardTitle,
+  CardDescription,
+  CardContent,
+  CardFooter,
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectScrollDownButton,
+  SelectScrollUpButton,
+  SelectSeparator,
+  SelectTrigger,
+  SelectValue,
+  Tabs,
+  TabsList,
+  TabsTrigger,
+  TabsContent,
+  Alert,
+  AlertTitle,
+  AlertDescription,
+  type AlertVariant,
+  type AlertProps,
+  Accordion,
+  AccordionItem,
+  AccordionTrigger,
+  AccordionContent,
+  RadioGroup,
+  RadioGroupItem,
+  ToggleGroup,
+  ToggleGroupItem,
+  Breadcrumb,
+  BreadcrumbList,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+  BreadcrumbEllipsis,
+  NavItem,
+  type NavItemProps,
+  NavLink,
+  type NavLinkProps,
+  NavGroup,
+  type NavGroupProps,
+} from './molecules'
+
+// Navigation and command organisms
+export {
+  Navigation,
+  NavigationMenu,
+  NavigationList,
+  NavigationItem,
+  NavigationTrigger,
+  NavigationContent,
+  NavigationLink,
+  NavigationBrand,
+  NavigationSeparator,
+  NavigationSpacer,
+  NavigationMobileToggle,
+  useNavigationDropdown,
+  Sidebar,
+  SidebarHeader,
+  SidebarContent,
+  SidebarFooter,
+  SidebarNav,
+  SidebarSection,
+  SidebarSeparator,
+  SidebarToggle,
+  Command,
+  CommandInput,
+  CommandList,
+  CommandEmpty,
+  CommandGroup,
+  CommandItem,
+  CommandSeparator,
+  CommandShortcut,
+  useCommandShortcut,
+} from './organisms'
+
+export type { SidebarItem, SidebarProps } from './organisms'
